+++ src/App.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import { Route } from 'react-router-dom';
+import { Route, Switch } from 'react-router-dom';
 import Menu from './components/Menu';
 import Story from './components/Story/Story';
 import Formation from './components/Formation/Formation';
@@ -11,11 +11,13 @@ class App extends Component {
 	render() {
 		return (
 			<div className="App">
-				<Route exact={true} path="/" component={Menu} />
-				<Route path="/story/" component={Story} />
-				<Route path="/formation/" component={Formation} />
-				<Route path="/characters/" component={Characters} />
-				<Route path="/account/" component={Account} />
+				<Switch>
+					<Route exact={true} path="/" component={Menu} />
+					<Route path="/story/" component={Story} />
+					<Route path="/formation/" component={Formation} />
+					<Route path="/characters/" component={Characters} />
+					<Route path="/account/" component={Account} />
+				</Switch>
 			</div>
 		);
 	}